ui/ops: consolidate trend ticker formatting and fix negative zero display ... Co-authored-by: Junie <junie@jetbrains.com>
