+++ ui/ops/src/dynamic/widgets/energy/PowerFactorCard.vue
@@ -50,6 +50,7 @@ import CircularGauge from '@/components/CircularGauge.vue';
 import {useRollingValue} from '@/composables/rollingValue.js';
 import {useElectricDemand, usePullElectricDemand} from '@/traits/electricDemand/electric.js';
 import {format} from '@/util/number.js';
+import {formatTrend, getTrendColor, getTrendIcon} from '@/util/trend.js';
 import {computed, toRef, watch} from 'vue';
 
 const props = defineProps({
@@ -147,46 +148,6 @@ const activeThreshold = computed(() => {
 const gaugeColor = computed(() => activeThreshold.value?.color ?? 'primary');
 const pfLabel = computed(() => activeThreshold.value?.label ?? 'No data');
 const pfIcon = computed(() => activeThreshold.value?.icon ?? 'mdi-help-circle-outline');
-
-/**
- * Get trend icon based on change value
- *
- * @param {number|null} change
- * @return {string} Material Design icon name
- */
-function getTrendIcon(change) {
-  if (change === null || change === undefined || Math.abs(change) < 0.05) return 'mdi-minus';
-  return change > 0 ? 'mdi-trending-up' : 'mdi-trending-down';
-}
-
-/**
- * Get trend color based on change value
- *
- * @param {number|null} change
- * @return {string} Vuetify color name
- */
-function getTrendColor(change) {
-  if (change === null || change === undefined || Math.abs(change) < 0.05) return 'grey-lighten-1';
-  return change > 0 ? 'error' : 'success';
-}
-
-/**
- * Format trend value for display
- *
- * @param {number|null} change
- * @return {string}
- */
-function formatTrend(change) {
-  if (change === null || change === undefined || Math.abs(change) < 0.05) return '0%';
-  const absChange = Math.abs(change);
-  if (absChange < 10) {
-    const formatted = change.toFixed(1);
-    if (formatted === '0.0' || formatted === '-0.0') return '0%';
-    return `${change > 0 ? '+' : ''}${formatted}%`;
-  }
-  const rounded = Math.round(change);
-  return `${rounded > 0 ? '+' : ''}${rounded}%`;
-}
 
 </script>
